app/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.9
gunicorn==21.2.0
packaging==23.2
psycopg2==2.9.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
Pillow==10.2.0
drf-spectacular==0.27.0
djangorestframework==3.14.0
django-filter==23.5
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
celery==5.3.6
redis==5.0.1
flower==1.2.0
django_celery_beat==2.5.0
django-json-widget==1.1.1
